add separate config for typescript files
