+++ package.json
@@ -50,6 +50,7 @@
     "format:fix": "prettier --write ."
   },
   "dependencies": {
+    "@eslint/js": "^9.4.0",
     "eslint-config-prettier": "^9.1.0",
     "globals": "^15.3.0",
     "typescript-eslint": "^7.11.0"
